Extract query class for Subscription.
